👷 Add Dependabot reviewer workflow
This commit adds a new workflow file `review-dependabot.yml` that sets up a Dependabot reviewer for pull requests. The workflow is triggered on the `pull_request_target` event and grants the necessary permissions for writing pull requests and contents. The workflow runs on `ubuntu-latest` and includes a step to review Dependabot PRs using the `pwshrc/actions-review-dependabot` action. The action requires the `GITHUB_TOKEN` secret and specifies that only pull requests by the user `dependabot[bot]` are required.